make folder for containing all IR related things, and move src/backend/ir.c/h to src/backend/ir/ssa.c/h
