+++ src/backend/codegen/nasm_x86_64_linux.h
@@ -1,7 +1,7 @@
 #ifndef NASM_X86_64_LINUX_H
 #define NASM_X86_64_LINUX_H
 
-#include "../ir.h"
+#include "../ir/ssa.h"
 #include <stdio.h>
 
 bool nasm_x86_64_linux_generate_file(FILE* sink, const IRModule* mod);
